% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_me.R
\name{about_me}
\alias{about_me}
\title{Tells you something about yourself}
\usage{
about_me(my_name, birth_year, fav_animal)
}
\arguments{
\item{my_name}{a character containing the input name}

\item{birth_year}{a number containing a four digit birth year}

\item{fav_animal}{a character containing an animal species}
}
\value{

}
\description{
Contains one function, about_me, that prints a statement about the user
}
