% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_vertebrates_doc.R
\docType{data}
\name{and_vertebrates}
\alias{and_vertebrates}
\title{Mack Creek aquatic vertebrates}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32209 rows and 18 columns.
}
\source{
{Enter dataset citation here}
\url{Add link for data package here}
}
\usage{
and_vertebrates
}
\description{
Sizes for trout and salamanders in Mack Creek (1987 - present)
}
\keyword{datasets}
