% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday}
\usage{
birthday(name, month, day)
}
\arguments{
\item{name}{The person's name (string)}

\item{month}{The person's birth month (numeric)}

\item{day}{The person's birth day (numeric)}
}
\value{

}
\description{
This function will introduce someone's name and state when their birthday is. Returns a statement
}
\examples{
birthday("Bob", 5, 22)
}
