% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(activity, milage)
}
\arguments{
\item{activity}{The activity done (past tense)}

\item{milage}{How many miles completed (numeric)}
}
\value{

}
\description{
Returns a statement stating the distance you did an activity
}
\examples{
distance("walked", 3)
}
