% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_even.R
\name{odd_even}
\alias{odd_even}
\title{Odd_Even}
\usage{
odd_even(num1, num2)
}
\arguments{
\item{num1}{The first number}

\item{num2}{The second number}
}
\value{

}
\description{
Adds two numbers together and says whether the sum is odd or even
}
\examples{
odd_even(10, 53)
}
